requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
